post-merge fixes
